requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
anthropic==0.21.0
Pillow==10.2.0
